ワイバーンの作成

+++ JapanGameGrandprize_2-4A/JapanGameGrandprize_2-4A/DragonBullet.cpp
@@ -21,8 +21,8 @@ DragonBullet::DragonBullet(Location spawn_location, Location player_location)
 	damage = BULLET_DAMAGE;
 	float radian; //Šp“x
 	radian = atan2f(player_location.y - location.y, player_location.x - location.x);
-	x_speed = speed * cosf(radian);
-	y_speed = speed * sinf(radian);
+	x_speed = static_cast<int>(speed * cosf(radian));
+	y_speed = static_cast<int>(speed * sinf(radian));
 }
 
 //-----------------------------------
